attempt to fix export
